Using extension layer

+++ function/lambda_function.py
@@ -10,7 +10,8 @@ import zipfile
 import tempfile
 
 from extutil import remove_none_attributes, gen_log, creturn, handle_common_errors, \
-    account_context, component_safe_name, ExtensionHandler, ext, lambda_env, random_id
+    account_context, component_safe_name, ExtensionHandler, ext, lambda_env, \
+    random_id, create_zip
 
 eh = ExtensionHandler()
 ALLOWED_RUNTIMES = ["python3.9", "python3.8", "python3.6", "python3.7", "nodejs14.x", "nodejs12.x", "nodejs10.x", "ruby2.7", "ruby2.5"]
@@ -80,6 +81,8 @@ def lambda_handler(event, context):
             eh.add_op("gen_props")
             if cdef.get("requirements"):
                 eh.add_op("add_requirements", cdef.get("requirements"))
+            elif cdef.get("requirements.txt"):
+                eh.add_op("add_requirements", "$$file")
         elif op == "delete":
             eh.add_op('remove_role')
             eh.add_op("remove_old", {"name": function_name})
@@ -239,16 +242,19 @@ def add_requirements(bucket, object_name):
             archive.extractall()
 
         requirements_file = f"{install_directory}requirements.txt"
-        with open(requirements_file, "w") as f:
-            f.writelines("%s\n" % i for i in requirements)
+        if requirements != "$$file":
+            with open(requirements_file, "w") as f:
+                f.writelines("%s\n" % i for i in requirements)
         
-        # if os.path.exists(requirements_file):
-        #     with open(requirements_file, "r") as f:
-        #         lines = list(f.readlines())
-        #         eh.add_log("Requirements Found", {"lines": lines})
+        if os.path.exists(requirements_file):
+            # with open(requirements_file, "r") as f:
+            #     lines = list(f.readlines())
+            #     eh.add_log("Requirements Found", {"lines": lines})
             
-        subprocess.call('pip install -r requirements.txt -t .'.split(), stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
-        eh.add_log("Requirements installed", {"requirements": requirements})
+            subprocess.call('pip install -r requirements.txt -t .'.split(), stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
+            eh.add_log("Requirements installed", {"requirements": requirements})
+        else:
+            eh.add_log("No Requirements to Install", {"files": os.listdir()})
 
         print(os.listdir())
 
@@ -432,12 +438,3 @@ def remove_function():
             eh.retry_error(str(e), 90 if create_and_delete else 15)
             eh.add_log(f"Error Deleting Function", {"name": function_to_delete}, True)
 
-def create_zip(file_name, path):
-    ziph=zipfile.ZipFile(file_name, 'w', zipfile.ZIP_DEFLATED)
-    # ziph is zipfile handle
-    for root, dirs, files in os.walk(path):
-        for file in files:
-            ziph.write(os.path.join(root, file), 
-                       os.path.relpath(os.path.join(root, file), 
-                                       os.path.join(path, '')))
-    ziph.close()
